delete-school functionality

+++ src/main/java/com/dougblizz/schooladmin/repositories/SchoolRepository.java
@@ -8,5 +8,5 @@ import java.io.Serializable;
 
 @Repository("schoolRepository")
 public interface SchoolRepository extends JpaRepository<School, Serializable> {
-    public abstract School findById(int id);
+    public abstract School findById(long id);
 }
